+++ CHANGELOG.md
@@ -2407,3 +2407,4 @@
 - [2026-06-29 09:29] perf(core): apply RepaintBoundary on animated core widgets
 - [2026-06-29 11:07] perf(auth): apply RepaintBoundary on animated auth widgets
 - [2026-06-29 12:59] feat(analytics): add bi-directional event bus handler for analytics
+- [2026-06-29 13:00] style(analytics): harmonize light and dark theme palette for analytics